modify middleware

+++ server/controllers/SubjectController.go
@@ -63,10 +63,10 @@ func (sc *SubjectController) Update(c *fiber.Ctx) error {
 
 	var subjectDTO dtos.SubjectDTO
 
-	// currentUser := c.Locals("user").(*ent.User)
-	// if !guards.IsAdmin(currentUser) {
-	// 	return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "Not authorized"})
-	// }
+	currentUser := c.Locals("user").(*ent.User)
+	if !guards.IsAdmin(currentUser) {
+		return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "Not authorized"})
+	}
 
 	if err := c.BodyParser(&subjectDTO); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid request"})
@@ -97,10 +97,10 @@ func (sc *SubjectController) Delete(c *fiber.Ctx) error {
 	if err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid ID"})
 	}
-	// currentUser := c.Locals("user").(*ent.User)
-	// if !guards.IsAdmin(currentUser) {
-	// 	return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "Not authorized"})
-	// }
+	currentUser := c.Locals("user").(*ent.User)
+	if !guards.IsAdmin(currentUser) {
+		return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "Not authorized"})
+	}
 
 	errDelete := sc.subjectService.Delete(id.String())
